+++ ChangeLog.md
@@ -1,6 +1,14 @@
 # Change Log
 
-## v0.5.5
+## v0.5.6
 
+- Now we can parse following command lines correctly:
+    - ``` export OPENSSL_INCLUDE_DIR=`brew --prefix openssl`/include ```
+    - ``` echo "`date` and `go version`" ```
+    - ``` echo `date` and `go version` ```
+
+## v0.5.5 - 2017-06-22
+
+- Added `make` completion.
 - Added `ssh` completion.
 - Fixed an alias extension issue.